ndk: Rework media error types
- Clean up the names: we're already in the `ndk` crate so no need to
  repeat that in the type name;
- Make the types available outside the `media` module, so that they can
  be used in `AMidi` bindings which are not located inside
  `libmediandk.so` (but `libamidi.so`);
- Add documentation comments;
- Guard `dead_code` helper functions with the right `feature` bound.
